# models/community.go
package models

import "time"

type Community struct {
	ID   int64  `json:"id" db:"community_id"`
	Name string `json:"name" db:"community_name"`
}
type CommunityDetail struct {
	ID        int64     `json:"id" db:"community_id"`
	Name      string    `json:"name" db:"community_name"`
	Intro     string    `json:"introduction omitempty" db:"introduction"`
	CreatTime time.Time `json:"create_time" db:"create_time"`
}